+++ src/main/scala/utils/Solver.scala
@@ -7,6 +7,7 @@
 package utils
 
 import edu.ie3.datamodel.graph.DistanceWeightedEdge
+import edu.ie3.osmogrid.exception.SolverException
 import edu.ie3.osmogrid.graph.OsmGraph
 import edu.ie3.util.osm.model.OsmEntity.Node
 import org.slf4j.LoggerFactory
@@ -223,23 +224,29 @@ object Solver {
           // finds the two connections needed to add the neighbor to the copied graph
           val connectionA = connections.getConnection(source, neighbor)
           val connectionB = connections.getConnection(neighbor, target)
-          copy.addConnection(connectionA)
-          copy.addConnection(connectionB)
 
-          // calculating the added weight
-          val addedWeight: Quantity[Length] = connectionA.distance
-            .add(connectionB.distance)
-            .subtract(removedEdge.getDistance)
+          (connectionA, connectionB) match {
+            case (Some(conA), Some(conB)) =>
+              copy.addConnection(conA)
+              copy.addConnection(conB)
 
-          Some(
-            StepResultOption(
-              copy,
-              neighbor,
-              List(connectionA, connectionB),
-              removedEdge,
-              addedWeight
-            )
-          )
+              // calculating the added weight
+              val addedWeight: Quantity[Length] = conA.distance
+                .add(conB.distance)
+                .subtract(removedEdge.getDistance)
+
+              Some(
+                StepResultOption(
+                  copy,
+                  neighbor,
+                  List(conA, conB),
+                  removedEdge,
+                  addedWeight
+                )
+              )
+            case _ =>
+              None
+          }
         }
       }.toList
   }
@@ -284,7 +291,14 @@ object Solver {
     )
 
     // connecting nodeA and nodeB
-    graph.addConnection(connections.getConnection(nodeA, nodeB))
+
+    connections.getConnection(nodeA, nodeB) match {
+      case Some(connection) => graph.addConnection(connection)
+      case None =>
+        throw SolverException(
+          s"An connection between $nodeA and $nodeB was expected"
+        )
+    }
 
     StepResult(graph, nodeA, nodes.diff(List(nodeToHv, nodeA, nodeB)))
   }